rename to filesHelper, add multiple product test
