+++ src/FilesHelper.ts
